+++ docs_src/src/guide/changelog.md
@@ -2,6 +2,11 @@
 
 [[toc]]
 
+## v2.2.0
+
+- Added [testing API](./advanced/testing.md).
+- Fixed links in documentation.
+
 ## v2.1.0
 
 - `filesystem.GetMIMEType` now detects `css`, `js`, `json` and `jsonld` files based on their extension.
@@ -21,7 +26,7 @@
 - Added [request data accessors](./basics/requests.html#accessors).
 - Some refactoring and package renaming have been done to better respect the Go conventions.
     - The `helpers` package have been renamed to `helper`
-- The server now shuts down when it encounters an error.
+- The server now shuts down when it encounters an error during startup.
 - New [`validation.GetFieldType`](./basics/validation.html#validation-getfieldtype) function.
 - Config and Lang are now protected with a `sync.RWMutex` to avoid data races in multi-threaded environments.
 - Greatly improve concurrency.